Move CharacterDetailsColors to core module
