+++ app/src/main/java/com/pnow/ramlist/app/ui/screen/character/DetailsHeroView.kt
@@ -20,9 +20,9 @@ import androidx.compose.ui.unit.sp
 import com.pnow.domain.model.CharacterStatus
 import com.pnow.ramlist.R
 import com.pnow.ramlist.app.ui.model.DetailsInfo
-import com.pnow.ramlist.app.ui.screen.details.CharacterDetailsColors
 import com.pnow.ramlist.app.ui.screen.previewDetails
 import com.pnow.ramlist.app.ui.screen.status.StatusPill
+import com.pnow.ramlist.core.ui.theme.CharacterDetailsColors
 import com.pnow.ramlist.core.ui.theme.RickAndMortyTheme
 
 @Composable
