--- app/src/main/java/com/pnow/ramlist/app/ui/screen/details/CharacterDetailsColors.kt
@@ -1,20 +1,0 @@
-package com.pnow.ramlist.app.ui.screen.details
-
-import androidx.compose.ui.graphics.Color
-
-object CharacterDetailsColors {
-    val Background = Color(0xFFF9F0EE)
-    val HeroBackground = Color(0xFFEDD8D3)
-    val CardBackground = Color(0xFFFFFFFF)
-
-    val TextDefault = Color(0xFF000000)
-    val TextPrimary = Color(0xFF2C1A17)
-    val TextMuted = Color(0xFF9B7B74)
-    val StatusPillBackground = Color(0xFF2C1A17)
-    val StatusPillText = Color(0xFFF9F0EE)
-    val StatusAlive = Color(0xFF5FCF7A)
-    val StatusDead = Color(0xFFE24B4A)
-    val EpisodeBadgeBackground = Color(0xFFF9F0EE)
-    val Error = Color(0xFFE24B4A)
-    val AvatarBackground = Color(0xFFD4B8B2)
-}
