+++ app/src/main/java/com/pnow/ramlist/app/ui/screen/details/CharacterDetailsScreen.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -25,6 +24,7 @@ import com.pnow.ramlist.app.ui.state.CharacterInfoState
 import com.pnow.ramlist.app.ui.state.DetailsUiState
 import com.pnow.ramlist.app.ui.state.EpisodeState
 import com.pnow.ramlist.app.ui.viewmodel.DetailsViewModel
+import com.pnow.ramlist.core.ui.theme.CharacterDetailsColors
 import com.pnow.ramlist.core.ui.theme.RickAndMortyTheme
 
 @Composable
@@ -47,8 +47,7 @@ private fun CharacterDetailsContent(state: DetailsUiState, onBackClick: () -> Un
         modifier =
         modifier
             .fillMaxSize()
-            .background(CharacterDetailsColors.Background)
-            .systemBarsPadding(),
+            .background(CharacterDetailsColors.Background),
     ) {
         item {
             DetailsHeroSection(
